requirements.txt
Flask==2.3.0
Werkzeug==2.3.0
requests==2.31.0
Pillow==11.2.1
flask-cors==4.0.0
pytest==7.4.0
pytest-asyncio==0.21.1